Move `MsgString` type into a separate module
